add URL Query parameters utils
