+++ src/api/buildListingEndpoint/index.ts
@@ -4,16 +4,7 @@ import {
   QueryParameter,
   BuildListingEndpointParameters,
 } from './models';
-
-const toRawQueryParameter = ({ name, value }): string => {
-  return `${name}=${JSON.stringify(value)}`;
-};
-
-const toRawQueryParameters = (queryParameters): Array<string> =>
-  queryParameters
-    .filter(({ value }) => value !== undefined && value.length !== 0)
-    .map(toRawQueryParameter)
-    .join('&');
+import toRawQueryParameters from '../../queryParameters';
 
 const getQueryParameters = ({
   sort,
