+++ src/index.ts
@@ -114,14 +114,20 @@ export { default as buildListingEndpoint } from './api/buildListingEndpoint';
 export {
   Parameters as ListingParameters,
   BuildListingEndpointParameters,
-  RegexSearchParameter,
-  ListsSearchParameter as ListSearchParameter,
   SearchParameter,
   SearchMatch,
-  QueryParameter,
 } from './api/buildListingEndpoint/models';
 export { default as buildListingDecoder } from './api/buildListingDecoder';
 
 export { default as copyToClipboard } from './utils/copy';
 export { default as useIntersectionObserver } from './utils/useIntersectionObserver';
 export { default as ContentWithCircularLoading } from './ContentWithCircularProgress';
+export {
+  setUrlQueryParameters,
+  getUrlQueryParameters,
+} from './queryParameters/url';
+export {
+  QueryParameter,
+  RegexSearchParameter,
+  ListsSearchParameter as ListSearchParameter,
+} from './queryParameters/models';
