+++ src/queryParameters/index.ts
@@ -1,0 +1,11 @@
+const toRawQueryParameter = ({ name, value }): string => {
+  return `${name}=${JSON.stringify(value)}`;
+};
+
+const toRawQueryParameters = (queryParameters): Array<string> =>
+  queryParameters
+    .filter(({ value }) => value !== undefined && value.length !== 0)
+    .map(toRawQueryParameter)
+    .join('&');
+
+export default toRawQueryParameters;
